check bookmark and mark as bookmarked using server actions
