--- app/api/bookmarks/route.js
@@ -1,55 +1,0 @@
-import connectDB from '@/config/database';
-import User from '@/models/User';
-import { getSessionUser } from '@/utils/getSessionUser';
-
-export const dynamic = 'force-dynamic';
-
-// NOTE: here we need to send back a Content-Type: application/json response
-// header rather than a text/plain header.
-
-// GET /api/bookmarks
-// We no longer need this route handler as the page is a server component so can
-// query the db directly.
-
-export const POST = async (request) => {
-  try {
-    await connectDB();
-
-    const { propertyId } = await request.json();
-
-    const sessionUser = await getSessionUser();
-
-    if (!sessionUser || !sessionUser.userId) {
-      return new Response('User ID is required', { status: 401 });
-    }
-
-    const { userId } = sessionUser;
-
-    // Find user in database
-    const user = await User.findOne({ _id: userId });
-
-    // Check if property is bookmarked
-    let isBookmarked = user.bookmarks.includes(propertyId);
-
-    let message;
-
-    if (isBookmarked) {
-      // If already bookmarked, remove it
-      user.bookmarks.pull(propertyId);
-      message = 'Bookmark removed successfully';
-      isBookmarked = false;
-    } else {
-      // If not bookmarked, add it
-      user.bookmarks.push(propertyId);
-      message = 'Bookmark added successfully';
-      isBookmarked = true;
-    }
-
-    await user.save();
-
-    return Response.json({ message, isBookmarked });
-  } catch (error) {
-    console.log(error);
-    return new Response('Something went wrong', { status: 500 });
-  }
-};
