+++ components/BookmarkButton.jsx
@@ -3,6 +3,8 @@ import { useState, useEffect } from 'react';
 import { useSession } from 'next-auth/react';
 import { toast } from 'react-toastify';
 import { FaBookmark } from 'react-icons/fa';
+import checkBookmarkStatus from '@/app/actions/checkBookmarkStatus';
+import bookmarkProperty from '@/app/actions/bookmarkProperty';
 
 const BookmarkButton = ({ property }) => {
   const { data: session } = useSession();
@@ -17,31 +19,14 @@ const BookmarkButton = ({ property }) => {
       return;
     }
 
-    const checkBookmarkStatus = async () => {
-      try {
-        const res = await fetch('/api/bookmarks/check', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            propertyId: property._id,
-          }),
-        });
-
-        if (res.status === 200) {
-          const data = await res.json();
-          setIsBookmarked(data.isBookmarked);
-        }
-      } catch (error) {
-        console.log(error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    checkBookmarkStatus();
-  }, [property._id, userId]);
+    // NOTE: here we can use a server action to check the bookmark status for a
+    // specific use for the property.
+    checkBookmarkStatus(property._id).then((res) => {
+      if (res.error) toast.error(res.error);
+      if (res.isBookmarked) setIsBookmarked(res.isBookmarked);
+      setLoading(false);
+    });
+  }, [property._id, userId, checkBookmarkStatus]);
 
   const handleClick = async () => {
     if (!userId) {
@@ -49,26 +34,13 @@ const BookmarkButton = ({ property }) => {
       return;
     }
 
-    try {
-      const res = await fetch('/api/bookmarks', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          propertyId: property._id,
-        }),
-      });
-
-      if (res.status === 200) {
-        const data = await res.json();
-        toast.success(data.message);
-        setIsBookmarked(data.isBookmarked);
-      }
-    } catch (error) {
-      console.log(error);
-      toast.error('Something went wrong');
-    }
+    // NOTE: here we can use a server action to mark bookmark a property for the
+    // user.
+    bookmarkProperty(property._id).then((res) => {
+      if (res.error) return toast.error(res.error);
+      setIsBookmarked(res.isBookmarked);
+      toast.success(res.message);
+    });
   };
 
   if (loading) return <p className='text-center'>Loading...</p>;
